+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The public API of this library consists of the functions declared in file
 ## [Unreleased]
 ### Added
 - CMake options for excluding filter applications or benchmarks from the build. (#247)
+### Changed
+- Argument parsing for all filter applications is more flexible. (#238)
 
 ## [3.4.4] - 2019-05-30
 ### Changed